requirements.txt
requests
beautifulsoup4
ics
google-genai
python-dotenv